fast-api/수집수집/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
beautifulsoup4==4.12.3
pymongo==4.8.0
python-dotenv==1.0.1
scikit-learn==1.5.2
wordcloud==1.9.3
matplotlib==3.9.2
apscheduler==3.10.4
